[bg-164361614] add handle error on JSON response
