+++ src/controllers/userMessages/getMessages.js
@@ -2,7 +2,7 @@ import messages from '../../model/messages';
 // import contacts from '../../model/contacts';
 
 export const GetMessages = (req, res, next) => {
-  const select = messages.filter(message => message.receiverId === parseInt(req.params.contactId, 10));
+  const select = messages.find(message => message.receiverId === parseInt(req.params.contactId, 10));
   if (select) {
     return res.status(200).send({
       status: 200,
@@ -12,6 +12,6 @@ export const GetMessages = (req, res, next) => {
 
   return res.status(404).send({
     status: 404,
-    error: `The contact was ${req.params.contactId} not found`
+    error: `The message for the contact with id ${req.params.contactId} is not found`
   });
 };
